Give the mmap ranges to handle close() on Drop.
This will make it easier for the handle to implement its trait, by no
longer requiring it to do some B-tree magic or linear search, to figure
out what it actually allocated. Instead, it receives an iterator (based
on the MMAP map, but ignoring not-yet-usable or leaked ranges), that the
closer can use to deallocate everything more easily.